Restructure cmd/server
